Ruby Slim: Fix highlighting for `p= some_ruby_code` lines.
The whole line had the scope `meta.tag`, thus “normal” text was displayed in the same color as the tag itself. (The highlighting was correct though if there was an additional space: `p = some_ruby_code`. But we have tons of Slim files without this space.)

The only change is in the end-regexp for `meta.tag`:
  - original: `$|(?!\.|#|=|:|-|~|/|\}|\]|\*|\s?[\*\{])`
  - vhanged: `$|(?!\.|#|:|-|~|/|\}|\]|\*|\s?[\*\{])`

I’m pretty sure that this change breaks some other (rare?) cases, but for now it looks fine …